Put the LockFile in the MainDir by default
Previous the default location of the `LockFile` was in `/var/run/`, but
the nzb user does not have permission to create a `LockFile` file in
that directory. With previous version of nzbget (ie. v13.0) this wasn't
really a problem because the service would still start up, but as of at
least v17.1 this is no longer the case and the service fails to start.

Placing the `LockFile` in `MainDir` will avoid this permission problem
and matches the default nzbget config as well.