Set correct Type in systemd service file
The systemd service file previously had `Type=simple` set, but nzbget is
started in daemon mode which means that the process forks. This was
causing the service to be killed as soon as it was started because
systemd didn't realise that it was forking.

Setting `Type=forking` resolves this problem, but we also need to set
the `PIDFile` so that systemd can keep track of the nzbget service
process.